requirements.txt
aiokafka==0.8.0
async-timeout==4.0.2
grpcio==1.51.1
grpcio-tools==1.51.1
kafka-python==2.0.2
packaging==22.0
protobuf==4.21.12
